fix(frontend): allow disconnect if account level error
