+++ packages/frontend/components/Shared/Header/Header.tsx
@@ -14,7 +14,6 @@ import {
   Fade,
   Button,
   Chip,
-  CircularProgress,
   Stack,
   Divider,
   ListItemText,
@@ -24,16 +23,13 @@ import Image from "next/image"
 import { useRouter } from "next/router"
 import { shallow } from "zustand/shallow"
 
-import { BurgerMenuIcon } from "./Icons"
-import ChainSelect from "./ChainSelect"
-import Parameters from "./Parameters"
-import styles from "../../styles/components/Header.module.css"
-import { Balances } from "@web3-onboard/core/dist/types"
-import AccountModal from "./AccountModal"
-import { useHistory } from "../../store/history.store"
-import Balance from "./Balance"
-import ParameterLinks from "./ParameterLinks"
-import { useAuth } from "../../store/auth.store"
+import { BurgerMenuIcon } from "../Icons"
+import ChainSelect from "../ChainSelect"
+import Parameters from "../Parameters"
+import styles from "../../../styles/components/Header.module.css"
+import ParameterLinks from "../ParameterLinks"
+import { useAuth } from "../../../store/auth.store"
+import BalanceAddress from "./BalanceAddress"
 
 const pages = [
   { name: "Markets", path: "/markets" },
@@ -72,7 +68,7 @@ const Header = () => {
 
   const formattedAddress =
     address?.substring(0, 5) + "..." + address?.substring(address?.length - 4)
-
+  // window.alert(balance)
   return (
     <AppBar position="static">
       <Box
@@ -271,74 +267,3 @@ const Header = () => {
   )
 }
 export default Header
-
-type BalanceAddressProps = {
-  address: string
-  formattedAddress: string
-  balance?: Balances
-  ens?: string
-}
-const BalanceAddress = (props: BalanceAddressProps) => {
-  const { palette } = useTheme()
-  const active = useHistory((state) => state.ongoingTxns.length)
-
-  const [accountModalEl, setAccountModalEl] = useState<
-    HTMLElement | undefined
-  >()
-  const showAccountModal = Boolean(accountModalEl)
-  const { balance, address, formattedAddress, ens } = props
-
-  if (!balance) {
-    return <></>
-  }
-
-  const [bal] = Object.values<string>(balance)
-  const [token] = Object.keys(balance)
-
-  const formattedBalance = <Balance balance={+bal} symbol={token} />
-  const pending = active && (
-    <Grid container alignItems="center">
-      <CircularProgress size={16} sx={{ mr: "0.625rem" }} />
-      <Typography
-        variant="small"
-        onClick={(e) => setAccountModalEl(e.currentTarget)}
-      >
-        {active} pending
-      </Typography>
-    </Grid>
-  )
-
-  return (
-    <Box mr="-2rem">
-      <Chip
-        label={formattedBalance}
-        sx={{ paddingRight: "2rem", fontSize: "0.875rem" }}
-      />
-      <Chip
-        onClick={(e) => setAccountModalEl(e.currentTarget)}
-        label={pending || ens || formattedAddress}
-        sx={{
-          background: palette.secondary.light,
-          borderRadius: "4rem",
-          height: "2.25rem",
-          padding: "0.438rem 0.75rem",
-          cursor: "pointer",
-          fontSize: "0.875rem",
-          position: "relative",
-          left: "-2rem",
-          backgroundColor: "#3C3D41", // Not part of the design system, one time use
-          border: `1px solid ${palette.secondary.light}`,
-          "&:hover": {
-            backgroundColor: palette.secondary.main,
-          },
-        }}
-      />
-      <AccountModal
-        isOpen={showAccountModal}
-        anchorEl={accountModalEl as HTMLElement}
-        closeAccountModal={() => setAccountModalEl(undefined)}
-        address={address}
-      />
-    </Box>
-  )
-}
