+++ packages/frontend/pages/index.tsx
@@ -3,7 +3,7 @@ import Head from "next/head"
 import Image from "next/image"
 
 import Footer from "../components/Shared/Footer"
-import Header from "../components/Shared/Header"
+import Header from "../components/Shared/Header/Header"
 import styles from "../styles/Home.module.css"
 
 const Home: NextPage = () => {
