+++ packages/frontend/pages/theming.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from "next"
 import Head from "next/head"
 import Theming from "../components/Theming/Theming"
-import Header from "../components/Shared/Header"
+import Header from "../components/Shared/Header/Header"
 
 const ThemingPage: NextPage = () => {
   return (
